Moved analyze_timing.py to utils.
